+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Here is a summary of the notable changes in the "rpm-changes" extension.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how
 to structure this file.
 
+## [Unreleased]
+
+- Refresh the "X minutes ago" hover messages every minute so they always use the
+  correct value, the others are refreshed once an hour.
+
 ## [1.1.0] - 2025-10-18
 
 - Validate the order of the changes, propose a quick fix (previous + 1 minute)